Add code for MVP
This adds code for a minimum working bot. Currently, it consists of just
one command, /test, that only runs in a specific test guild.

The important thing is that I'm laying a good foundation by separating
commands out into their own cogs and using configuration files instead
of hardcoding values in the code.